requirements.txt
# Env requirments
pip
autopep8

# Application Requirments
fastapi
uvicorn
pymongo
pydantic
cryptography
starlette
pyjwt